Title: Homography and object recognition v2
Short: I made a feature were the user can calculate the homography of an detected object.
Detail: Using the homography the user can calculate things like
- position
- perspective
- depth
- orientation
- etc

Notice message: If you get an error like "floating-point-exception-core-dump",
this is because of the frame memory of opencv when using a webcam. just unplug the device and re run the code a couple of times

Changes: Changed opencv_type.hpp to use 'using' instead of typedef
Fixes: fixed opencv install guide

Author: Arsalan Anwari
Team: Scrum team 6
Date: 06-06-2018